+++ LNAndroidLibs/LNAndroidLibs/src/main/java/com/lnkj/libs/core/Any.kt
@@ -1,0 +1,9 @@
+package com.lnkj.libs.core
+
+//var Any.msg: String
+//    set(value) = value
+//    get()
+
+fun Any.setMsg(msg: String){
+
+}