+++ LNAndroidLibs/LNAndroidLibs/src/main/java/com/lnkj/libs/net/Result.kt
@@ -1,0 +1,7 @@
+package com.lnkj.libs.net
+
+data class Result<T>(
+    val msg: String,
+    val code: Int,
+    val data: T
+)